+++ src/main/kotlin/org/sert2521/infiniterecharge2020/Main.kt
@@ -1,10 +1,12 @@
 package org.sert2521.infiniterecharge2020
 
 import kotlinx.coroutines.launch
+import org.sert2521.infiniterecharge2020.drivetrain.Drivetrain
 import org.sert2521.sertain.robot
+import org.sert2521.sertain.subsystems.add
 
 suspend fun main() = robot {
+    add<Drivetrain>()
     launch {
-        println("Hello")
     }
 }
