+++ src/main/kotlin/org/sert2521/infiniterecharge2020/OI.kt
@@ -1,9 +1,8 @@
-package org.sert2521.infiniterecharge2020.oi
+package org.sert2521.infiniterecharge2020
 
 import edu.wpi.first.wpilibj.Joystick
 import edu.wpi.first.wpilibj.XboxController
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser
-import org.sert2521.infiniterecharge2020.Operator
 
 
 enum class ControlMode {
@@ -17,7 +16,6 @@ val controlModeChooser = SendableChooser<ControlMode>().apply {
 
 val controlMode get() = controlModeChooser.selected ?: ControlMode.JOYSTICK
 
+val primaryController by lazy { XboxController(Operator.PRIMARY_CONTROLLER) }
 val primaryJoystick by lazy { Joystick(Operator.PRIMARY_STICK) }
 val secondaryJoystick by lazy { Joystick(Operator.SECONDARY_STICK) }
-val primaryController by lazy { XboxController(Operator.PRIMARY_CONTROLLER) }
-
